refactor: move to `Consensus/State`
