+++ Libplanet.Net/Consensus/State/PreVoteState.cs
@@ -2,7 +2,7 @@ using Libplanet.Action;
 using Libplanet.Consensus;
 using Libplanet.Net.Messages;
 
-namespace Libplanet.Net.Consensus
+namespace Libplanet.Net.Consensus.State
 {
     public class PreVoteState<T> : IState<T>
         where T : IAction, new()
@@ -35,26 +35,7 @@ namespace Libplanet.Net.Consensus
                 throw new UnexpectedBlockHashException(vote);
             }
 
-            RoundContext<T> roundContext = context.CurrentRoundContext;
-
-            if (context.NodeId == vote.NodeId &&
-                vote.BlockHash.Equals(roundContext.BlockHash) &&
-                !context.ContainsBlock(vote.BlockHash))
-            {
-                context.VoteHolding.Set();
-                throw new VoteBlockNotExistsException(vote);
-            }
-
-            if (context.NodeId == vote.NodeId &&
-                roundContext.CurrentNodeVoteFlag is VoteFlag.Null &&
-                vote.BlockHash.Equals(roundContext.BlockHash) &&
-                context.ContainsBlock(vote.BlockHash))
-            {
-                context.VoteHolding.Reset();
-                roundContext.Vote(vote.ProposeVote);
-                return new ConsensusVote(roundContext.Voting(VoteFlag.Absent));
-            }
-
+            RoundContext<T> roundContext = context.RoundContextOf(vote.Height, vote.Round);
             roundContext.Vote(vote.ProposeVote);
 
             if (!roundContext.VoteSet.HasTwoThirdPrevote())
